Remove cast

+++ src/main/java/x1/jboss/syslog/SyslogHandler.java
@@ -203,7 +203,7 @@ public class SyslogHandler extends Handler {
   public void setFacility(String facility) {
     if (Syslog.FACILITY_MAP.containsKey(facility)) {
       this.facility = facility;
-      this.facilityInt = (Integer) Syslog.FACILITY_MAP.get(facility);
+      this.facilityInt = Syslog.FACILITY_MAP.get(facility);
     } else {
       throw new IllegalArgumentException("Illegal syslog facility name: " + facility);
     }
